Created UserFavorites publication, did some tests for AllMenuitesm
Unfortunately, the plan I had for AllMenuItems failed (only publishing the MenuItems that are owned by a vendor/admin so the MenuItems will be unique - to "favorite" a MenuItem, a user will essentially create a copy of the MenuItem with them as the owner instead, but we don't want that to be displayed in the Food Listings page. Going to try create a master user that will have a copy of each item added by vendors/another admin to be displayed as the AllMenuItems.